Split blackcat av from binds module
