+++ witch_craft/src/modules/blackcat/mod.rs
@@ -1,0 +1,1 @@
+pub mod blackcat;
