+++ witch_craft/src/modules/mod.rs
@@ -1,4 +1,5 @@
 pub mod binds;
+pub mod blackcat;
 pub mod core;
 pub mod llm;
 pub mod network;
